Ajusta geração de logs e feedback da importação

+++ dspacedashboard/imports/views.py
@@ -9,69 +9,85 @@ from django.shortcuts import render
 from django.http import HttpResponseRedirect
 from django.views.generic.base import TemplateView
 from django.core.files.storage import FileSystemStorage
+from django.views.generic import DetailView
 from django.views.generic.edit import FormView
 from django.views.generic import ListView
+from django.shortcuts import render
 
 from dspacedashboard.imports.forms import ImportFileForm
 from dspacedashboard.imports.models import FileImport, Collection
+from dspacedashboard.core.dspace_utils import get_collections
 
-class ImportFileView(FormView):
+
+def import_file(request):
     template_name = 'imports/import_file.html'
-    success_message = 'Importação concluída com sucesso'
-    form_class = ImportFileForm
+    success_message = 'Importação realizada com sucesso'
 
-    def dispatch(self, *args, **kwargs):
-        try:
-            response = requests.get('http://localhost:8080/solr/search/select?q=search.resourcetype:3&fl=dc.title,handle&wt=json&omitHeader=true&rows=5000')
-            self.collections = json.loads(response.text)['response']['docs'] if response.status_code == 200 else []
-        except Exception as e:
-            print("Error: ", e)
-            self.collections = [] 
-        return super(ImportFileView, self).dispatch(*args, **kwargs)
+    collections = get_collections()
+    collections_form = ()
+    for collection in collections:               
+        if collection.get('dc.title', None):
+            collections_form += ((collection['handle'], collection['dc.title'][0]),)
+    
+    form = ImportFileForm(request.POST or None, request.FILES or None, collections=collections_form)
+    context = {}
 
-    def get_form_kwargs(self):
-        kwargs = super(ImportFileView, self).get_form_kwargs()
-
-        collections = ()
-        for collection in self.collections:                
-            if collection.get('dc.title', None):
-                collections = collections + ((collection['handle'], collection['dc.title'][0]),)
-
-        kwargs.update({
-		    'collections' : collections
-		})
-        return kwargs
-
-    def get_success_url(self):		
-        return reverse('import:file')
-
-    def form_valid(self, form):
+    if form.is_valid():
         #Getting/save target collection and file data
         handle = form.cleaned_data.get('collection')        
-        collection_name = next(item for item in self.collections if item["handle"] == handle)
+        collection_name = next(item for item in collections if item["handle"] == handle)
         collection, created = Collection.objects.get_or_create(handle=handle, name=collection_name.get('dc.title', [['']])[0])
-        file_import = FileImport.objects.create(user=self.request.user, collection=collection)
+        file_import = FileImport.objects.create(user=request.user, collection=collection)
         
         #Saving temporary file
         fs = FileSystemStorage()
         upload_file = form.files.get('dspace_simple_archive', None)
         filename = fs.save(upload_file.name, upload_file)
-        
-        #Setting logger up
-        logging.basicConfig(filename=f'log/{str(file_import.id)}.log', level=logging.INFO)
 
         unziped_dir = os.path.join(settings.MEDIA_ROOT, str(file_import.id))
-        output = subprocess.check_output(['unzip', fs.path(filename), '-d', unziped_dir])
+        output = subprocess.check_output(['unzip', fs.path(filename), '-d', unziped_dir])         
+        
+        import_dir = os.path.join(settings.MEDIA_ROOT, unziped_dir, os.listdir(unziped_dir)[0])
+        import_dir = os.path.join(import_dir, os.listdir(import_dir)[0])
 
-        logging.info(output.decode("utf-8"))
+        #COMANDO DE IMPORTAÇÃO AQUI   
+        output += subprocess.check_output(['ping', '-c', '10', '127.0.0.1'])
 
         fs.delete(filename)
-        shutil.rmtree(unziped_dir)
+        shutil.rmtree(unziped_dir)     
+        
+        logger = open(f'log/{str(file_import.id)}.log', 'w+')
+        for line in output.splitlines():            
+            logger.write(line.decode('utf-8'))
+            logger.write('\n')
+                    
+        context['import_output'] = output.decode('utf-8').splitlines()
 
-        messages.info(self.request, self.success_message)		
-        return HttpResponseRedirect(self.get_success_url())
+        messages.info(request, success_message)
+        logger.close()
+
+    context['form'] = form
+    return render(request, template_name, context)
 
 
 class ImportFileListView(ListView):
     template_name = 'imports/import_list.html'
     model = FileImport
+
+
+class ImportLogDetailView(DetailView):
+    template_name = 'imports/import_log.html'
+    model = FileImport
+
+    def get_context_data(self, **kwargs):
+        context = super(ImportLogDetailView, self).get_context_data(**kwargs)
+        try:
+            with open(f'log/{str(self.object.id)}.log', 'r') as log:  
+                lines = []
+                for line in log:
+                    lines.append(line)
+            context['log_content'] = lines
+            log.close()
+        except Exception as e:
+            context['log_content'] = [e]
+        return context
